+++ src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { CommingSoonComponent, UnauthorizedComponent } from './modules/admin-shared';
-import { HomeComponent } from './home/home.component';
+import { HomeComponent } from './modules/admin-core/components/home/home.component';
 import { UsersGuard, RouteGuard } from './modules/admin-core';
