+++ src/app/app.module.ts
@@ -5,7 +5,7 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { CoreModule, TranslateService, SharedModule } from 'shikshalokam';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { HeaderComponent } from './header/header.component';
+import { HeaderComponent } from './modules/admin-core/components/header/header.component';
 import { LoginComponent } from './login/login.component';
 import { JwtModule } from '@auth0/angular-jwt';
 import { AdminCoreModule } from './modules/admin-core/admin-core.module';
@@ -13,9 +13,9 @@ import { keyCloakService } from './modules/admin-core';
 import { AdminSharedModule } from './modules/admin-shared';
 import { environment } from 'src/environments/environment';
 import { KeycloakAngularModule } from 'keycloak-angular';
-import { HomeComponent } from './home/home.component';
-import { BreadCrumComponent } from './bread-crum/bread-crum.component';
-import { AdminFooterComponent } from './admin-footer/admin-footer.component';
+import { HomeComponent } from './modules/admin-core/components/home/home.component';
+import { BreadCrumComponent } from './modules/admin-core/components/bread-crum/bread-crum.component';
+import { AdminFooterComponent } from './modules/admin-core/components/admin-footer/admin-footer.component';
 
 
 export function tokenGetter() {
