--- src/app/bread-crum/bread-crum.component.scss
@@ -1,4 +1,0 @@
-@import "../../assets/scss/variables.scss";
-.negative {
-    color: $primary-color;
-  }
