+++ src/app/modules/admin-core/components/admin-footer/admin-footer.component.scss
@@ -1,4 +1,4 @@
-@import '../../assets/scss/variables.scss';
+@import "../../../../../assets/scss/variables.scss";
 .footer {
   height: 40px;
   // border-top:1px solid var(--light-grey-color);
