+++ src/app/modules/admin-core/components/admin-footer/admin-footer.component.ts
