+++ src/app/modules/admin-core/components/bread-crum/bread-crum.component.scss
@@ -1,0 +1,4 @@
+@import "../../../../../assets/scss/variables.scss";
+.negative {
+    color: $primary-color;
+  }
