+++ src/app/modules/admin-core/components/header/header.component.scss
@@ -1,4 +1,4 @@
-@import "../../assets/scss/variables.scss"; 
+@import "../../../../../assets/scss/variables.scss";
 .space {
     padding: 15px;
 }
