+++ src/app/modules/admin-core/components/home/home.component.scss
@@ -1,4 +1,4 @@
-@import "../../assets/scss/variables.scss";
+@import "../../../../../assets/scss/variables.scss";
 
 .carddata {
   width: 360px;
