+++ src/styles.scss
@@ -2817,3 +2817,7 @@ html.cdk-global-scrollblock body {
 .cdk-overlay-container {
   overflow: auto;
 }
+
+.p-button{
+  background: #bababa;
+}
